Test separate tsconfig
